+++ src/lib.rs
@@ -29,7 +29,7 @@
 //! }
 //!
 //! serializer! {
-//!     serialize_user: User {
+//!     serialize_user<User> {
 //!         attr(id)
 //!         attr(name)
 //!         has_one(country, serialize_country)
@@ -38,7 +38,7 @@
 //! }
 //!
 //! serializer! {
-//!     serialize_country: Country {
+//!     serialize_country<Country> {
 //!         attr(id)
 //!     }
 //! }
@@ -95,7 +95,7 @@
 //! # }
 //! #
 //! # serializer! {
-//! #     serialize_country: Country {
+//! #     serialize_country<Country> {
 //! #         attr(id)
 //! #     }
 //! # }