+++ src/lib.rs
@@ -126,7 +126,7 @@
     unused_import_braces,
     unused_qualifications
 )]
-#![doc(html_root_url = "https://docs.rs/serializers/0.1.1")]
+#![doc(html_root_url = "https://docs.rs/serializers/0.1.2")]
 
 extern crate serde;
 #[macro_use]