+++ CHANGELOG.md
@@ -16,7 +16,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ### Removed
 
-- Removed `ToJson` trait. Its could be merged with `Serializer`.
+N/A
 
 ### Fixed